refactor: moved to new utils file
