+++ src/utils/index.ts
@@ -1,5 +1,6 @@
 import * as FormatUtils from "./format.utils";
 import * as ImageUtils from "./image.utils";
+import * as JSONUtils from "./JSON.utils";
 import * as LocalStorageUtils from "./localstorage.utils";
 import * as ObjectUtils from "./object.utils";
 import * as PromiseUtils from "./promise.utils";
@@ -10,6 +11,7 @@ import * as OSDetectionUtils from "./os-detect.utils";
 
 export {
     ImageUtils,
+    JSONUtils,
     FormatUtils,
     LocalStorageUtils,
     ObjectUtils,
